後ろの列追加
うしろの列のアニメーション処理を追加
Scoreの引数の型修正
Needleにアニメーション切替のタイミング用変数追加

+++ GameJam2025/GameJam/Object/Score.cpp
@@ -46,7 +46,7 @@ void Score::Finalize()
 
 }
 
-void Score::AddScore(float red_y, float event_y)
+void Score::AddScore(int red_y, int event_y)
 {
 	diff = std::abs(red_y - event_y);
 	Evaluate();
